refactor: 金額と数量の計算をBigDecimalに寄せる
- 金額、数量、損益計算の型を Double から BigDecimal に変更し、計算時の近似誤差が出にくいように修正
- json状態の互換性を担保するため、専用の BigDecimalSerializer を作成
- BigDecimal による安全な演算（RoundingMode付きの divide など）に SimulationService を適応し、関連する Repository と Test も改修

+++ projects/crypto-autotrading-app/src/main/kotlin/cryptoautotrading/domain/repository/TradeHistoryRepository.kt
@@ -1,5 +1,7 @@
 package cryptoautotrading.domain.repository
 
+import java.math.BigDecimal
+
 /**
  * 取引履歴を保存するインターフェース
  */
@@ -9,11 +11,11 @@ interface TradeHistoryRepository {
      */
     fun append(
         datetime: String,
-        price: Double,
+        price: BigDecimal,
         sign: String,
         reason: String,
-        profitAndLoss: Double,
+        profitAndLoss: BigDecimal,
         isHolding: Boolean,
-        fee: Double
+        fee: BigDecimal
     )
 }
